:recycle: refactor internals
