+++ docs/v2/parsing/error.rst
@@ -1,0 +1,15 @@
+V2 Parsing Error
+################
+
+
+Error Item
+----------
+.. autoclass:: mindee.v2.parsing.error.error_item.ErrorItem
+    :members:
+    :inherited-members:
+
+Error Response
+--------------
+.. autoclass:: mindee.v2.parsing.error.error_response.ErrorResponse
+    :members:
+    :inherited-members:
