+++ docs/v2/parsing/index.rst
@@ -6,3 +6,5 @@ V2 Parsing
     :maxdepth: 3
 
     ./inference/index
+    ./error
+    ./job
