+++ docs/v2/parsing/job.rst
@@ -1,0 +1,21 @@
+V2 Parsing Job
+##############
+
+
+Job
+---
+.. autoclass:: mindee.v2.parsing.job.job.Job
+    :members:
+    :inherited-members:
+
+Job Response
+------------
+.. autoclass:: mindee.v2.parsing.job.job_response.JobResponse
+    :members:
+    :inherited-members:
+
+Job Webook
+----------
+.. autoclass:: mindee.v2.parsing.job.job_webhook.JobWebhook
+    :members:
+    :inherited-members:
