+++ mindee/v2/parsing/__init__.py
@@ -1,11 +1,11 @@
+from mindee.v2.parsing.error.error_item import ErrorItem
+from mindee.v2.parsing.error.error_response import ErrorResponse
 from mindee.v2.parsing.inference.base_inference import BaseInference
 from mindee.v2.parsing.inference.base_response import BaseResponse
-from mindee.v2.parsing.inference.error_item import ErrorItem
-from mindee.v2.parsing.inference.error_response import ErrorResponse
 from mindee.v2.parsing.inference.inference_active_options import InferenceActiveOptions
 from mindee.v2.parsing.inference.inference_file import InferenceFile
 from mindee.v2.parsing.inference.inference_model import InferenceModel
-from mindee.v2.parsing.inference.job_response import JobResponse
+from mindee.v2.parsing.job.job_response import JobResponse
 from mindee.v2.product.extraction.extraction_inference import ExtractionInference
 from mindee.v2.product.extraction.extraction_response import ExtractionResponse
 from mindee.v2.product.extraction.extraction_result import ExtractionResult
