+++ mindee/v2/parsing/error/__init__.py
@@ -1,0 +1,4 @@
+from mindee.v2.parsing.error.error_item import ErrorItem
+from mindee.v2.parsing.error.error_response import ErrorResponse
+
+__all__ = ["ErrorItem", "ErrorResponse"]
