+++ mindee/v2/parsing/job/__init__.py
@@ -1,0 +1,5 @@
+from mindee.v2.parsing.job.job import Job
+from mindee.v2.parsing.job.job_response import JobResponse
+from mindee.v2.parsing.job.job_webhook import JobWebhook
+
+__all__ = ["Job", "JobResponse", "JobWebhook"]
